test: split pr checks graph assertions under the line ratchet
release_workflow_config_test grew past 500 lines when the rust-test
needs lock was added. move that suite to its own file.
